chore(login-page): separating front-end logic on login page
